Fix for putting summary in slide template
In my testing, I hit an issue when rendering the template due to a
missing function "raw" for an issue. Upon further investigation, simply
accessing the "summary" field for the issue worked just fine.

It is curious that others have not hit this issue, so care should be
taken in ensuring that this change does not break others' use
cases/workflows.