finishline: Add service-account-password
Another team has asked for the ability to use a service account via
basic auth. To do so they will need to pass in a password. This change
provides a new switch which can be used to pass in a password for use
with basic auth. The switch contains a warning/reminder that passwords
stored in a config file or via a switch are viewable and the name itself
is meant to be a reminder that it should only be used in service account
type situations.

Signed-off-by: Steve Milner <smilner@redhat.com>