Hide FACTORY-1946 for factory reports
This issue is one of our Objective epics. It's causing issues in
reporting due to the fact that it contains issues instead of exclusively
epics. The code expects for the epic to have a parent epic, but since
this is a top level epic is doesn't. The code fails on looking for the
parent epic for this epic.

This issue could be fixed by ignoring the failure when the parent epic
relationship doesn't work, however I believe the current behavior is
probably more correct, and that the correct thing to do is hide the
epic.