Use mosaic to calculate epic progress
The factory team has begun working on epics to increase the team's
velocity. These epics have specific goals around increasing or
decreasing a key metric by X percent. We'd like to report on progress on
these epics by calculating the current metric number as a percentage of
the goal.

We've developed a tool, mosaic, that calculates the values of the key
metrics we wish to track on. This tool can be found at [1]

When generating reports using finishline, we want to have the progress
percentage for these epics calculated using mosaic. This change enables
this functionality. I've added a new argument '--mosaic-config' to the
finshline script. When this argument is present, finishline will parse
the specified config file to determine which issues to use mosaic for,
as well as to determine how to calculate progress. If the argument is
not specified, finishline behaves as normal.

[1] https://github.com/accorvin/jira-mosaic