Explicitly handle arguments as unicode
I encountered an issue running finishline with Python 3 where execution
failed on decoding a string. This change implements what I believe is
the correct fix for this, specifically:

  - Explicitly convert all command line string arguments as utf-8
  unicode
  - Convert these strings to unicode in a way that will work for both
  python 2 and python 3 (no conversion is necessary for python 3)

Note the change from ascii to utf-8 when writing the output file. I'm
not sure if there was a reason for requiring ascii, but I encountered a
formatting issue on the final output version when running in python 3
and forcing utf-8 seemed to fix it.