Use additional authentication path for Kerberos
Internally our client needs to authenticate against a separate authentication
path first in order to communicate with JIRA. The current code is not doing this
and therefore will not authenticate properly even if the user who runs the
script actually has a ticket.

This patch allows to configure an authentication path and unconditionally
disables mutual authentication for Kerberos.